+++ NEWS.md
@@ -1,9 +1,7 @@
 **Version ?.?.? - ?**
 
 - Added option ADAGUC_TRACE_TIMINGS to measure the amount spent on db access, file reading and image generation.
-
-**Version 3.1.2. 2025-06-10**
-
+- GenericDataWarper has been refactored.
 - Prevent invalid filenames in tests
 
 **Version 3.1.1. 2025-06-05**